move cached_method to _statements.py
